+++ changelog.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Upcoming
+### Changes
+* Reduced Explosive enchantment cost
+
 ## 1.20.9
 ### Changes
 * Hang glider (by Fuzs) recipe is now made in forge